crew_avaliacao/requirements.txt
fastapi
uvicorn
crewai
python-dotenv